Move quat helpers
